# Check consistency of constraints before evaluating a pattern