--- a/booster/library/Booster/JsonRpc.hs
+++ b/booster/library/Booster/JsonRpc.hs
@@ -166,6 +166,7 @@ respond stateVar request =
 
                         case evaluatedInitialPattern of
                             (Left ApplyEquations.SideConditionFalse{}, _) -> do
+                                -- input pattern's constrains are Bottom, return Bottom
                                 stop <- liftIO $ getTime Monotonic
                                 pure $
                                     execResponse