… booster/test/rpc-integration/test-log-simplify-json/simplify-log.txt.golden
… booster/test/rpc-integration/test-substitutions/README.md
… booster/test/rpc-integration/test-substitutions/response-circular-equations.booster-dev
… booster/test/rpc-integration/test-substitutions/response-symbolic-bottom-predicate.booster-dev
… booster/test/rpc-integration/test-substitutions/response-symbolic-bottom-predicate.json
… booster/test/rpc-integration/test-substitutions/response-symbolic-two-substitutions.booster-dev
… booster/test/rpc-integration/test-vacuous/README.md
… booster/test/rpc-integration/test-vacuous/response-vacuous-but-rewritten.json
… booster/test/rpc-integration/test-vacuous/response-vacuous-not-rewritten.json
… booster/test/rpc-integration/test-vacuous/response-vacuous-not-rewritten.kore-rpc-dev
… booster/test/rpc-integration/test-vacuous/response-vacuous-without-rewrite.json
… booster/test/rpc-integration/test-vacuous/state-vacuous-not-rewritten.execute
… booster/unit-tests/Test/Booster/Pattern/Rewrite.hs
… dev-tools/package.yaml
… dev-tools/pretty/Pretty.hs
… scripts/booster-integration-tests.sh
… scripts/performance-tests-kevm.sh