Fixes #4012